class-32/demo/blogapi-postgress/pyproject.toml
[tool.poetry]
name = "blogapi-postgress"
version = "0.1.0"
description = ""
authors = ["Ahmad Alawad <ahmad.alawad.sf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2.5"
djangorestframework = "^3.12.4"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
